feat: integrate Gemini AI service for vehicle spec generation

+++ backend/autospec-api/src/main/java/br/com/autospec/backend/dto/VehicleResponseDTO.java
@@ -5,6 +5,6 @@ public record VehicleResponseDTO(
         int horsepower,
         int torque,
         String drivetrain,
-        int price
+        double price
 ) {
 }
